Move various tests from ``TestTypeChecker`` to functional tests (#5455)
* Remove duplicate test already in 'not_context_manager.py'
