+++ tests/checkers/unittest_typecheck.py
@@ -200,179 +200,6 @@ class TestTypeChecker(CheckerTestCase):
         with self.assertAddsMessages(message):
             self.checker.visit_attribute(node)
 
-    @set_config(
-        contextmanager_decorators=(
-            "contextlib.contextmanager",
-            ".custom_contextmanager",
-        )
-    )
-    def test_custom_context_manager(self):
-        """Test that @custom_contextmanager is recognized as configured."""
-        node = astroid.extract_node(
-            """
-        from contextlib import contextmanager
-        def custom_contextmanager(f):
-            return contextmanager(f)
-        @custom_contextmanager
-        def dec():
-            yield
-        with dec():
-            pass
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_with(node)
-
-    def test_invalid_metaclass(self) -> None:
-        module = astroid.parse(
-            """
-        class InvalidAsMetaclass(object):
-            pass
-
-        class FirstInvalid(object, metaclass=int):
-            pass
-
-        class SecondInvalid(object, metaclass=InvalidAsMetaclass):
-            pass
-
-        class ThirdInvalid(object, metaclass=2):
-            pass
-
-        class FourthInvalid(object, metaclass=InvalidAsMetaclass()):
-            pass
-        """
-        )
-        for class_obj, metaclass_name in (
-            ("FourthInvalid", "Instance of .InvalidAsMetaclass"),
-            ("ThirdInvalid", "2"),
-            ("SecondInvalid", "InvalidAsMetaclass"),
-            ("FirstInvalid", "int"),
-        ):
-            classdef = module[class_obj]
-            message = MessageTest(
-                "invalid-metaclass", node=classdef, args=(metaclass_name,)
-            )
-            with self.assertAddsMessages(message):
-                self.checker.visit_classdef(classdef)
-
-    def test_invalid_metaclass_function_metaclasses(self) -> None:
-        module = astroid.parse(
-            """
-        def invalid_metaclass_1(name, bases, attrs):
-            return int
-        def invalid_metaclass_2(name, bases, attrs):
-            return 1
-        class Invalid(metaclass=invalid_metaclass_1):
-            pass
-        class InvalidSecond(metaclass=invalid_metaclass_2):
-            pass
-        """
-        )
-        for class_obj, metaclass_name in (("Invalid", "int"), ("InvalidSecond", "1")):
-            classdef = module[class_obj]
-            message = MessageTest(
-                "invalid-metaclass", node=classdef, args=(metaclass_name,)
-            )
-            with self.assertAddsMessages(message):
-                self.checker.visit_classdef(classdef)
-
-    def test_typing_namedtuple_not_callable_issue1295(self) -> None:
-        module = astroid.parse(
-            """
-        import typing
-        Named = typing.NamedTuple('Named', [('foo', int), ('bar', int)])
-        named = Named(1, 2)
-        """
-        )
-        call = module.body[-1].value
-        callables = call.func.inferred()
-        assert len(callables) == 1
-        assert callables[0].callable()
-        with self.assertNoMessages():
-            self.checker.visit_call(call)
-
-    def test_typing_namedtuple_unsubscriptable_object_issue1295(self) -> None:
-        module = astroid.parse(
-            """
-        import typing
-        MyType = typing.Tuple[str, str]
-        """
-        )
-        subscript = module.body[-1].value
-        with self.assertNoMessages():
-            self.checker.visit_subscript(subscript)
-
-    def test_staticmethod_multiprocessing_call(self) -> None:
-        """Make sure not-callable isn't raised for descriptors
-
-        astroid can't process descriptors correctly so
-        pylint needs to ignore not-callable for them
-        right now
-
-        Test for https://github.com/PyCQA/pylint/issues/1699
-        """
-        call = astroid.extract_node(
-            """
-        import multiprocessing
-        multiprocessing.current_process() #@
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_call(call)
-
-    def test_not_callable_uninferable_property(self) -> None:
-        """Make sure not-callable isn't raised for uninferable
-        properties
-        """
-        call = astroid.extract_node(
-            """
-        class A:
-            @property
-            def call(self):
-                return undefined
-
-        a = A()
-        a.call() #@
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_call(call)
-
-    def test_descriptor_call(self) -> None:
-        call = astroid.extract_node(
-            """
-        def func():
-            pass
-
-        class ADescriptor:
-            def __get__(self, instance, owner):
-                return func
-
-        class AggregateCls:
-            a = ADescriptor()
-
-        AggregateCls().a() #@
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_call(call)
-
-    def test_unknown_parent(self) -> None:
-        """Make sure the callable check does not crash when a node's parent
-        cannot be determined.
-        """
-        call = astroid.extract_node(
-            """
-        def get_num(n):
-            return 2 * n
-        get_num(10)()
-        """
-        )
-        with self.assertAddsMessages(
-            MessageTest("not-callable", node=call, args="get_num(10)")
-        ):
-            self.checker.visit_call(call)
-
 
 class TestTypeCheckerOnDecorators(CheckerTestCase):
     "Tests for pylint.checkers.typecheck on decorated functions."
