+++ tests/functional/n/not_callable.txt
@@ -1,9 +1,10 @@
-not-callable:5:0:5:10::REVISION is not callable:UNDEFINED
-not-callable:23:12:23:22::INSTANCE is not callable:UNDEFINED
-not-callable:25:12:25:18::LIST is not callable:UNDEFINED
-not-callable:27:12:27:18::DICT is not callable:UNDEFINED
-not-callable:29:12:29:19::TUPLE is not callable:UNDEFINED
-not-callable:31:12:31:17::INT is not callable:UNDEFINED
-not-callable:66:0:66:13::PROP.test is not callable:UNDEFINED
-not-callable:67:0:67:13::PROP.custom is not callable:UNDEFINED
-not-callable:135:18:135:45::ClassWithProperty().value is not callable:UNDEFINED
+not-callable:6:0:6:10::REVISION is not callable:UNDEFINED
+not-callable:24:12:24:22::INSTANCE is not callable:UNDEFINED
+not-callable:26:12:26:18::LIST is not callable:UNDEFINED
+not-callable:28:12:28:18::DICT is not callable:UNDEFINED
+not-callable:30:12:30:19::TUPLE is not callable:UNDEFINED
+not-callable:32:12:32:17::INT is not callable:UNDEFINED
+not-callable:67:0:67:13::PROP.test is not callable:UNDEFINED
+not-callable:68:0:68:13::PROP.custom is not callable:UNDEFINED
+not-callable:137:18:137:45::ClassWithProperty().value is not callable:UNDEFINED
+not-callable:190:0:190:16::get_number(10) is not callable:UNDEFINED
